requirements.txt
# Alphatic Portfolio Analyzer - Requirements
# Install with: pip install -r requirements.txt --break-system-packages

# Core Framework
streamlit>=1.28.0

# Data Analysis & Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Financial Data
yfinance>=0.2.28

# Statistics & Math
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Portfolio Analytics
pyfolio-reloaded>=0.9.5

# OpenBB Platform (for future features)
openbb>=4.0.0

# Optional: Performance Optimization
# numba>=0.57.0  # Uncomment if you want faster calculations
